Extract interface of input status service
